requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
pydantic==1.10.2
firebase-admin==5.3.0
python-dotenv==0.21.0
email-validator==1.3.0
pytest
pytest-asyncio
httpx